Removing Gemfile.lock to eliminate vuln
The version of rest-client this was built with has a vuln in it now. I'm not going to update this, but I might want the code, so I'm deleting the lock file so that no one gets the old, vulnerable version. This probably means that the code won't work anymore.